Extract report interpretation from PDF mapping
