+++ docs/ai/repo-map.md
@@ -38,7 +38,7 @@ Scope: single source of truth for detailed file layout, entry points, package st
 - `infra/config/`: runtime settings store (single `SettingsStore` owns both analysis and device settings) used by runtime wiring and exposed to use-cases through focused shared read ports.
 - `infra/workers/`: worker-pool infrastructure.
 - `use_cases/diagnostics/`: post-stop diagnostics pipeline. See `docs/analysis_pipeline.md` for the full data flow, pipeline steps, and module map. `findings.py` owns reference-check and peak/order orchestration, `peak_binning.py` owns persistent-peak accumulation/scoring, `signal_aggregation.py` owns speed/location summary aggregation, `math_utils.py` owns generic statistics/correlation helpers, and `speed_profile_helpers.py` owns speed-profile assembly helpers; `top_cause_selection.py` delegates classification and ranking to the domain `Finding`; `location_analysis.py` owns the location-analysis pipeline and `LocationAnalysisResult` typed return; `_types.py` is now limited to diagnostics-local helper types such as `AccelStatistics`, while boundary summary payloads live in `shared/boundaries/analysis_payload.py`. The package `__init__.py` re-exports only high-level analysis entrypoints; PDF rendering types/helpers now live under `adapters/pdf/`.
-- `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict used only inside history workflows.
+- `use_cases/history/`: run query/delete, PDF report generation, CSV/ZIP exports, and history-facing helper orchestration above persistence ports. `helpers.py` owns the local `HistoryRecord` TypedDict plus generic history workflow helpers, while `report_interpretation.py` owns pure report-domain interpretation consumed by the PDF mapper.
 - `use_cases/run/`: recording pipeline orchestration; `lifecycle_state.py` owns `RunLifecycleState` plus the active-run snapshot, `persistence_writer.py` owns history-write coordination/retry bookkeeping, `logger.py` owns `RunRecorder`, `post_analysis.py` owns the background analysis queue above the injected run-persistence/error boundary, and `sample_builder.py` owns pure sample-building helpers.
 - `use_cases/updates/`: wheel-based updater workflow orchestration, firmware cache, ESP flashing, release discovery, install, rollback, runner, Wi-Fi, and status tracking.
 - `shared/`: cross-cutting typed payloads (`shared/types/`), boundary serializers/decoders (`shared/boundaries/`), exceptions (`shared/exceptions.py`), JSON helpers (`shared/json_utils.py`), location identifiers (`shared/locations.py`), run-context helpers, and shared pure vehicle-order math (`shared/order_bands.py`). `shared/types/run_persistence.py` owns the current run-history persistence port used by history and recording use-cases, `shared/types/client_tracker.py` owns the active-client lookup port consumed by the recording path, `shared/types/signal_source.py` owns the recording path's latest-metrics/latest-samples port, `shared/types/speed_provider.py` owns the recording path's resolved-speed port, `shared/types/settings_reader.py` owns the read-only settings port consumed by recording and history flows, `shared/types/sensor_frame.py` owns the canonical typed sample-record shape used by recording and persistence, `shared/types/health_snapshot.py` owns runtime/persistence health snapshot TypedDicts shared by recording and HTTP health reporting, `shared/boundaries/run_log.py` owns JSONL run-log decoding/normalization shared by diagnostics and persistence, and `shared/boundaries/diagnostic_case.py` owns summary projection, typed speed/suitability decoding, and shared metadata-to-case reconstruction (`case_context_from_metadata`).
